rename global sentences.api.vm.ts to sentences.api.mappers.ts
