+++ src/core/sentences/index.ts
@@ -1,4 +1,4 @@
 export * from './global-sentences.model';
-export * from './global-sentences.api.vm';
+export * from './global-sentences.api.mappers';
 export * from './global-sentences.api';
 export * from './global-sentences.storage';
